pyproject.toml
[project]
name = "brickflow-cicd-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "brickflows>=1.3.1",
    "databricks>=0.2",
    "delta-spark>=3.3.0",
]
